# refactor: improve calculated fields aggregation