# Description

This PR splits the calculated fields aggregation step into two to optimize the performance. Even

## Useful links

- Please insert link to ticket: https://oortcloud.atlassian.net/jira/software/projects/HIT/boards/5?assignee=712020%3A4bfd0977-2d85-4a7e-973b-6c6731a06b1e&selectedIssue=HIT-442


## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
-  [x] Improvement (refactor or addition to existing functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update

# How Has This Been Tested?

* Created a form with a number field
* Create a lot of records for that field
* Create 10 calculated fields (`{{calc.sub( {{calc.add( {{calc.sub( {{calc.add( {{calc.sub( {{calc.add( {{calc.sub( {{calc.add( {{calc.sub( {{calc.add( {{calc.sub( {{calc.add( {{data.number}};1 )}};1 )}};1 )}};1 )}};1 )}};1 )}};1 )}};1 )}};1 )}};1 )}};1 )}};1 )}}`)
* Create a couple grids, selecting the number field and all the calculated fields
* On one grid, add a filter by only one of the calculated fields
* Measure request times, before and after change

### Before changes

| # | Without filter | With filter  |
|-------------|----------------|--------------|
| 1           | 414ms          | 395ms        |
| 2           | 415ms          | 398ms        |
| 3           | 564ms          | 400ms        |
| 4           | 389ms          | 383ms        |
| 5           | 428ms          | 397ms        |
| 6           | 406ms          | 397ms        |
| 7           | 377ms          | 392ms        |
| 8           | 391ms          | 390ms        |
| 9           | 387ms          | 474ms        |
| 10          | 385ms          | 371ms        |
| **avg** | **415.6ms**    | **399.7ms**  |

### After changes

| # | Without filter | With filter  |
|-------------|----------------|--------------|
| 1           | 373ms          | 390ms        |
| 2           | 372ms          | 389ms        |
| 3           | 364ms          | 366ms        |
| 4           | 364ms          | 490ms        |
| 5           | 375ms          | 370ms        |
| 6           | 377ms          | 364ms        |
| 7           | 367ms          | 386ms        |
| 8           | 348ms          | 360ms        |
| 9           | 444ms          | 391ms        |
| 10          | 375ms          | 369ms        |
| **avg** | **375.9ms**    | **387.5ms**  |








## Screenshots

Please include screenshots of this change. If this issue is only back-end related, and does not involve any visual change of the platform, you can skip this part.

# Checklist:

( \* == Mandatory )

- [ ] - I have set myself as assignee of the pull request
- [ ] - My code follows the style guidelines of this project
- [ ] - Linting does not generate new warnings
- [ ] - I have performed a self-review of my own code
- [ ] - I have put the ticket for review, adding the oort-backend team to the list of reviewers
- [ ] - I have commented my code, particularly in hard-to-understand areas
- [ ] - I have put JSDoc comment in all required places
- [ ] - My changes generate no new warnings
- [ ] - I have included screenshots describing my changes if relevant
- [ ] - I have selected labels in the Pull Request, according to the changes with code brings
- [ ] I have made corresponding changes to the documentation ( if required )
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream modules

# More explanation

https://www.loom.com/share/05a716d61b9744faaf51fb304c21d1e5?sid=f87cf896-582a-4f76-93ae-8ceed801b145